Move roundCost function to utilities
